销毁

+++ packages/lone-logic/component/index.js
@@ -47,10 +47,14 @@ class LogicComponent {
     callHook(vm, 'beforeDestroy')
     vm._isBeingDestroyed = true
     vm.data = null
-    instanceStorage.delete(vm._id)
+    vm._slave = null
     callHook(vm, 'destroyed')
-    vm._slave.send('component:destroy', this._id, {})
-    console.log('清除组件内所有事件绑定')
+    vm.$options = {}
+
+    if (instanceStorage.has(vm._id)) {
+      instanceStorage.delete(vm._id)
+    }
+    // vm._slave.send('component:destroy', this._id, {}) 销毁真实dom
   };
 }
